% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{finalise_plot_captionmagick}
\alias{finalise_plot_captionmagick}
\title{Arrange alignment and save ggplot chart}
\usage{
finalise_plot_captionmagick(
  plot_name,
  source_name,
  save_filepath = file.path(Sys.getenv("TMPDIR"), "tmp-nc.png"),
  width = 16,
  height = 9,
  logo_image_path = file.path(system.file("data", package = "zicolors"),
    "zilogo_square.png")
)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you want to format and save}

\item{source_name}{The text you want to in the bottom left hand side of your plot}

\item{save_filepath}{Exact filepath that you want the plot to be saved to}

\item{width}{Width in cm that you want to save your chart to - defaults to 16}

\item{height}{Height in cm that you want to save your chart to - defaults to 9}

\item{logo_image_path}{File path for the logo image you want to use in the right hand side of your chart,
which needs to be a PNG file - defaults to Zi square-ish logo that sits within the data folder of your package}
}
\value{
(Invisibly) an updated ggplot object.
}
\description{
Running this function will save your plot with left align your title, subtitle and right align source, add the Zi logo at the bottom right and save it to your specified location.
}
\examples{
\dontrun{
finalise_plot_captionmagick(plot_name = myplot,
source = "The source for my data",
save_filepath = "filename_that_my_plot_should_be_saved_to-nc.png",
width = 16,
height = 9,
logo_image_path = "logo_image_filepath.png"
)
}

}
\keyword{finalise_plot_captionmagick}
