% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scale_color_zi}
\alias{scale_color_zi}
\title{Zi color scale for ggplot2 with different pallettes}
\usage{
scale_color_zi(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{Options}{include palette (Defaults to "main" zi Blue/Green/Orange, with other options beeing based on mostly 2 colors except of "qualitative" with 4 colors)}
}
\description{
Zi color scale for ggplot2 with different pallettes
}
\examples{
\dontrun{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +   geom_point() +  theme_zi() + scale_color_zi()
}
}
\keyword{custom_scale_color()}
