% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scale_fill_zi}
\alias{scale_fill_zi}
\title{Zi fill scale for ggplot2 with different palettes}
\usage{
scale_fill_zi(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{Options}{include palette (Defaults to "main" zi Blue-Shades, with other options being "all", "qualitative")}
}
\description{
Zi fill scale for ggplot2 with different palettes
}
\examples{
\dontrun{
ggplot(iris, aes(x=Species,y=Sepal.Width, fill = Species)) +   geom_bar(stat = "summary", fun.y = "mean") +  theme_zi() + scale_fill_zi()
}
}
\keyword{custom_scale_color()}
