% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{theme_zi}
\alias{theme_zi}
\title{Zi Theme based on theme_grey}
\usage{
theme_zi(fontsize = 14, font = "Calibri")
}
\arguments{
\item{Define}{a base_size (Defaults to 12) and base_family and bold_family for Fonts used (defaults to ggplot2's defaults)}
}
\description{
Zi Theme based on theme_grey
}
\examples{
library(extrafont)
library(tidyverse)
font_import()
loadfonts(device="win")
ggplot(as.data.frame(Titanic) \%>\% group_by(Class) \%>\% summarise(n=sum(Freq)), aes(x=Class,  y=n)) + geom_bar(stat="identity" , fill=zi_cols("ziblue")) +  theme_zi() 
}
\keyword{theme}
