% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{theme_zi_horizontal}
\alias{theme_zi_horizontal}
\title{Zi Theme with for horizontal charts}
\usage{
theme_zi_horizontal(fontsize = 10, font = NULL)
}
\arguments{
\item{Define}{a base_size (Defaults to 12) and base_family for Fonts used (defaults to ggplot2's defaults)}
}
\description{
Zi Theme with for horizontal charts
}
\keyword{theme}
