% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{theme_zidatalab}
\alias{theme_zidatalab}
\title{Zi Datalab Theme with grey bg axis and labels based on theme_zi() optimal suited to print maps.}
\usage{
theme_zidatalab(fontsize = 14, font = "Calibri")
}
\arguments{
\item{Define}{a base_size (Defaults to 12) and base_family for Fonts used (defaults to ggplot2's defaults)}
}
\description{
Zi Datalab Theme with grey bg axis and labels based on theme_zi() optimal suited to print maps.
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +   geom_point() +  theme_zidatalab() +  scale_color_zi()
}
\keyword{theme}
